refactor: View, API 컨트롤러 분리
